Retain input volume to better simulate mute toggle
The output volume remains unchanged when muted and subsequently
un-muted.  Attempt to simulate this behavior by persisting the input
volume when mic-mute is toggled on and then using this persisted value
when toggled off.

If the user uses another method for muting the mic, e.g. navigating to
Sound under System Properties, then the last value of the input volume
will not be saved (but a prior value will still be in the property).
Note: only one value is saved, not a value for each input device.