Store separate volume per input source
The volume is stored separately for each input source such that, when
the source changes, the input volume does too.  If we store one volume
to retain for muting purposes, and the device switches, we may update
the new source to an inappropriate level on mic-mute toggle.  Avoid this
by storing a pre-mute volume _for each device_, and restoring that
volume when mute is disabled.